refactor: remove final dto single speed compressor train
