+++ src/libecalc/domain/process/compressor/dto/__init__.py
@@ -3,7 +3,6 @@ from .stage import CompressorStage, InterstagePressureControl
 from .train import (
     CompressorTrainSimplifiedWithKnownStages,
     CompressorTrainSimplifiedWithUnknownStages,
-    SingleSpeedCompressorTrain,
     VariableSpeedCompressorTrain,
     VariableSpeedCompressorTrainMultipleStreamsAndPressures,
 )
